Cambiamento tipo parametro page a int16 nella funzione RicercaLibro

+++ internal/db/database.go
@@ -363,7 +363,7 @@ func GetCurrentPrestito(codice uint32) (Prestito, error) {
 }
 
 //Funzione per la ricerca dei libri
-func RicercaLibri(nome string, autore, genere []uint32, page uint16) ([]Libro, error) {
+func RicercaLibri(nome string, autore, genere []uint32, page int16) ([]Libro, error) {
 	//Verifico se il server è ancora disponibile
 	//Se c'è un errore, ritorna null e l'errore
 	if err := db_Connection.Ping(); err != nil {
@@ -382,7 +382,7 @@ func RicercaLibri(nome string, autore, genere []uint32, page uint16) ([]Libro, e
 	for _, g := range genere {
 		args = append(args, g)
 	}
-	args = append(args, page*config.Config.Generale.LunghezzaPagina, (page+1)*config.Config.Generale.LunghezzaPagina)
+	args = append(args, uint16(page)*config.Config.Generale.LunghezzaPagina, uint16(page+1)*config.Config.Generale.LunghezzaPagina)
 
 	//Esamino tutti i casi possibili di richiesta, scegliendo la query giusta per ogni situazione possibile
 	q := `SELECT Libro.Codice,Titolo,Autore.Nome,Autore.Cognome,Genere.Nome,Prenotato,Hashz FROM Libro,Autore,Genere WHERE Libro.Autore = Autore.Codice AND Libro.Genere = Genere.Codice` + strings.Repeat(` AND titolo LIKE ?`, len(tags))
